# docs/openapi.yaml
openapi: 3.1.0
info:
  title: hushd API
  version: 0.7.0
  description: |
    hushd platform API: auth, verification, compliance (2257), billing, creator
    flows, media pipeline, messaging, payouts, moderation, and admin controls.

    ## Authentication

    Protected endpoints accept either:

    - `Authorization: Bearer <accessToken>` (API clients, tests), or
    - the `hushd_access` HttpOnly cookie set by `/auth/login`, `/auth/register`,
      `/auth/verify-email`, and `/auth/refresh` (browser frontends).

    Cookie-authenticated **mutating** requests must also send
    `X-Requested-With: XMLHttpRequest` (CSRF backstop). The refresh token
    travels in the `hushd_refresh` HttpOnly cookie scoped to `/v1/auth`;
    `/auth/refresh` and `/auth/logout` also accept it in the JSON body for
    non-browser clients.
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: http://localhost:3001/v1
    description: Local
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register account
      description: |
        Issues tokens and sets the auth cookies when the account activates
        immediately (dev bypass). Otherwise returns `pendingVerification` and
        sends a verification email — no session is created until the email
        token round-trips.
      operationId: authRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Tokens issued, or pending email verification
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AuthTokensResponse"
                  - $ref: "#/components/schemas/PendingVerificationResponse"
        "409":
          description: email_in_use
  /auth/login:
    post:
      tags: [auth]
      summary: Login account
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Tokens issued (auth cookies set)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokensResponse"
        "401":
          description: invalid_credentials | totp_required | invalid_totp
        "403":
          description: email_not_verified | account_not_active
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token
      description: |
        Reads the refresh token from the request body or, for browser clients,
        from the `hushd_refresh` HttpOnly cookie. Rotates the session and sets
        fresh cookies. Reuse of a revoked refresh token revokes all sessions
        for the user.
      operationId: authRefresh
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Tokens issued (auth cookies set)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokensResponse"
        "401":
          description: invalid_refresh | refresh_reuse
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke refresh token session
      description: Idempotent — always clears the auth cookies and returns ok.
      operationId: authLogout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Session revoked, cookies cleared
  /auth/verify-email:
    post:
      tags: [auth]
      summary: Consume an email verification token
      description: Activates a PENDING account and issues a session.
      operationId: authVerifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "200":
          description: Tokens issued (auth cookies set)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokensResponse"
        "400":
          description: invalid_or_expired_token
  /auth/verify-email/resend:
    post:
      tags: [auth]
      summary: Resend the verification email
      description: Always succeeds — the response never reveals whether the email exists.
      operationId: authResendVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailAddressRequest"
      responses:
        "200":
          description: ok
  /auth/password-reset/request:
    post:
      tags: [auth]
      summary: Request a password reset email
      description: Always succeeds — the response never reveals whether the email exists.
      operationId: authRequestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailAddressRequest"
      responses:
        "200":
          description: ok
  /auth/password-reset/confirm:
    post:
      tags: [auth]
      summary: Consume a password reset token
      description: Sets the new password and revokes all existing sessions.
      operationId: authConfirmPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirmRequest"
      responses:
        "200":
          description: Password updated
        "400":
          description: invalid_or_expired_token
  /auth/me:
    get:
      tags: [auth]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Current user profile
      operationId: authMe
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Not authenticated
  /auth/totp/setup:
    post:
      tags: [auth]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Generate a TOTP secret (creators only)
      operationId: authTotpSetup
      responses:
        "200":
          description: otpauth URL for enrolment
        "403":
          description: creators_only
  /auth/totp/enable:
    post:
      tags: [auth]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Enable TOTP with a verification code (creators only)
      operationId: authTotpEnable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpEnableRequest"
      responses:
        "200":
          description: TOTP enabled
        "400":
          description: totp_not_setup | totp_already_enabled
        "401":
          description: invalid_totp
...
  /moderation/queue:
    get:
      tags: [moderation]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Moderation queue (moderator/admin)
      operationId: moderationQueue
      responses:
        "200":
          description: Queue items
...
  /admin/moderation/queue:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Moderation queue for admins
      operationId: adminModerationQueue
      responses:
        "200":
          description: Queue items
  /admin/moderation/queue/{id}/resolve:
    post:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Resolve a moderation queue item
      operationId: adminModerationResolve
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Resolved
  /admin/payouts:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: List payouts for review
      operationId: adminPayoutList
      responses:
        "200":
          description: Payout rows
  /admin/payouts/{payoutId}/approve:
    post:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Approve a pending payout
      operationId: adminPayoutApprove
      parameters:
        - $ref: "#/components/parameters/payoutId"
      responses:
        "200":
          description: Approved
  /admin/payouts/{payoutId}/paid:
    post:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Mark an approved payout as paid
      operationId: adminPayoutPaid
      parameters:
        - $ref: "#/components/parameters/payoutId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                processorRef: { type: string }
      responses:
        "200":
          description: Marked paid
  /admin/payouts/{payoutId}/cancel:
    post:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Cancel a payout
      operationId: adminPayoutCancel
      parameters:
        - $ref: "#/components/parameters/payoutId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: Canceled
  /admin/risk/holds:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: List active risk holds
      operationId: adminRiskHolds
      responses:
        "200":
          description: Holds
  /admin/risk/holds/{holdId}/release:
    post:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Release a risk hold
      operationId: adminRiskHoldRelease
      parameters:
        - name: holdId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Released
  /admin/fraud/dashboard:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Fraud dashboard aggregates
      operationId: adminFraudDashboard
      responses:
        "200":
          description: Fraud summary
  /admin/compliance/export:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Export compliance records (2257)
      operationId: adminComplianceExport
      responses:
        "200":
          description: Export payload
  /admin/geo-blocks:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: List geo-block rules
      operationId: adminGeoBlockList
      responses:
        "200":
          description: Rules
    post:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Create a geo-block rule
      operationId: adminGeoBlockCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGeoBlockRequest"
      responses:
        "201":
          description: Rule created
  /admin/geo-blocks/{countryCode}:
    delete:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Delete geo-block rule(s) for a country
      operationId: adminGeoBlockDelete
      parameters:
        - name: countryCode
          in: path
          required: true
          schema: { type: string, pattern: "^[A-Za-z]{2}$" }
      responses:
        "200":
          description: Deleted
  /admin/dmca:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: List DMCA notices
      operationId: adminDmcaList
      responses:
        "200":
          description: Notices
  /admin/dmca/{id}/status:
    post:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Update DMCA notice status / legal hold
      operationId: adminDmcaStatus
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDmcaStatusRequest"
      responses:
        "200":
          description: Updated
  /admin/audit-logs:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      summary: Admin audit log
      operationId: adminAuditLogs
      responses:
        "200":
          description: Audit rows
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: hushd_access
      description: HttpOnly session cookie set by the auth endpoints.
  parameters:
...
    payoutId:
      name: payoutId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, maxLength: 72 }
        role: { type: string, enum: [FAN, CREATOR] }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
        totpCode: { type: string, pattern: '^\d{6}$', description: Required when TOTP is enabled }
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: Optional in the body — browser clients send it via the hushd_refresh cookie.
    EmailAddressRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token: { type: string }
    PasswordResetConfirmRequest:
      type: object
      required: [token, password]
      properties:
        token: { type: string }
        password: { type: string, minLength: 12, maxLength: 72 }
    TotpEnableRequest:
      type: object
      required: [code]
      properties:
        code: { type: string, pattern: '^\d{6}$' }
...
    CreateGeoBlockRequest:
      type: object
      required: [countryCode, scope]
      properties:
        countryCode: { type: string, pattern: "^[A-Za-z]{2}$" }
        scope: { type: string, enum: [ACCESS, PAYOUTS] }
        reason: { type: string, maxLength: 500 }
    UpdateDmcaStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [RECEIVED, UNDER_REVIEW, ACTIONED, REJECTED, CLOSED] }
        legalHold: { type: boolean }
    PendingVerificationResponse:
      type: object
      required: [pendingVerification, user]
      properties:
        pendingVerification: { type: boolean, enum: [true] }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
    MeResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        roles:
          type: array
          items: { type: string, enum: [FAN, CREATOR, ADMIN, MODERATOR] }
        status: { type: string, enum: [PENDING, ACTIVE, SUSPENDED, DELETED] }
        emailVerified: { type: boolean }
        createdAt: { type: string, format: date-time }
        creatorSlug: { type: string, nullable: true }
        totpEnabled: { type: boolean }
    AuthTokensResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn, user]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, enum: [Bearer] }
        expiresIn: { type: integer }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
            roles:
              type: array
              items: { type: string }